+++ src/features/core/components/PhotoDisplay.tsx
