+++ src/features/navbar/components/NavbarWrapper.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from "react-router-dom";
 import Navbar from "./Navbar";
-import Footer from "./Footer";
+import Footer from "../../core/components/Footer";
 
 export default function NavbarWrapper() {
   return (
